pyproject.toml
[tool.poetry]
name = "fast_api_test"
version = "0.1.0"
description = ""
authors = ["Gleb Garanin <glebgar567@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
pydantic = "^1.8.2"
email-validator = "^1.1.3"
Jinja2 = "^3.0.1"
PyYAML = "^5.4.1"
python-multipart = "^0.0.5"
graphene = "^3.0"
asyncpg = "^0.25.0"
environs = "^9.3.2"
uvloop = "^0.16.0"
SQLAlchemy = "^1.4.31"
aioredis = "^2.0.0"
passlib = "^1.7.4"
fastapi_jinja = { git = "https://github.com/AGeekInside/fastapi-jinja.git", branch = "main" }
sqlalchemy2-stubs = "^0.0.2a19"
Authlib = "^0.15.4"
itsdangerous = "^2.0.1"
argon2-cffi = "^21.3.0"
PyJWT = "^2.3.0"
aio-pika = "^6.8.1"
structlog = "^21.5.0"
structlog-sentry = "^1.4.0"
orjson = "^3.6.6"
colorlog = "^6.6.0"
uvicorn = "^0.17.0post1"
omegaconf = "^2.1.1"
cattrs = "^1.10.0"

[tool.poetry.dev-dependencies]
alembic = "^1.6.5"
pytest = "^6.2.4"
pytest-asyncio = "^0.17.2"
pytest-cov = "^3.0.0"
black = { version = "^21.12b0", python = ">=3.6" }
mypy = { extras = ["plugin"], version = "^0.931" }
timeout-decorator = "^0.5.0"
flake8 = "^4.0.1"
asynctest = "^0.13.0"
async-timeout = "^4.0.2"
pytest-aiohttp = "^1.0.3"
asgi-lifespan = "^1.0.1"
gunicorn = "^20.1.0"
httpx = "0.22.0"
pytest-mock = "^3.6.1"
nest-asyncio = "^1.5.1"
types-orjson = "^3.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
default = true